chore(physics): remove unused system/components
