+++ src/plugins/physics/src/components.rs
@@ -1,11 +1,11 @@
 pub(crate) mod affected;
+pub(crate) mod anchor;
 pub(crate) mod async_collider;
 pub(crate) mod blockable;
 pub(crate) mod blocker_types;
 pub(crate) mod collider;
 pub(crate) mod default_attributes;
 pub(crate) mod effects;
-pub(crate) mod fix_points;
 pub(crate) mod ground_target;
 pub(crate) mod hollow;
 pub(crate) mod interaction_target;
