+++ src/plugins/physics/src/components/mount_points.rs
@@ -1,4 +1,5 @@
 use bevy::prelude::*;
+use common::{tools::bone_name::BoneName, traits::thread_safe::ThreadSafe};
 use std::{collections::HashMap, hash::Hash};
 
 #[derive(Component, Debug, PartialEq)]
@@ -9,6 +10,21 @@ where
 impl<T> Default for MountPoints<T>
 where
 	T: Eq + Hash,
+{
+	fn default() -> Self {
+		Self(HashMap::default())
+	}
+}
+
+#[derive(Component, Debug, PartialEq, Clone)]
+#[require(MountPoints<T>)]
+pub struct MountPointsDefinition<T>(pub(crate) HashMap<BoneName, T>)
+where
+	T: Eq + Hash + ThreadSafe;
+
+impl<T> Default for MountPointsDefinition<T>
+where
+	T: Eq + Hash + ThreadSafe,
 {
 	fn default() -> Self {
 		Self(HashMap::default())
