+++ src/plugins/physics/src/systems.rs
@@ -1,8 +1,8 @@
+pub(crate) mod anchor_system;
 pub(crate) mod apply_pull;
 pub(crate) mod async_collider;
 pub(crate) mod blockable;
 pub(crate) mod body;
-pub(crate) mod fix_points;
 pub(crate) mod ground_target;
 pub(crate) mod insert_affected;
 pub(crate) mod interactions;
