--- src/plugins/physics/src/systems/fix_points.rs
@@ -1,2 +1,0 @@
-pub(crate) mod anchors;
-pub(crate) mod insert;
